format-html circularlity
